commit the change!

+++ library/src/main/java/zzg/klinechart/mycandle/entry/CandleEntry.java
@@ -8,14 +8,14 @@ public class CandleEntry {
     public   float  low;   //最低价
     public   float  open;  //开盘价
     public   float  close;  //收盘价
-    public   int    volume;   //成交量
-    public   int    volumeMoney;  //成交价
+    public   float    volume;   //成交量
+    public   float    volumeMoney;  //成交价
     public   String date;      //成交日期
     public   float xValue;   //平均价
     public   Boolean isRise;  //是否升降
 
 
-    public CandleEntry(float high, float low, float open, float close, int volume, int volumeMoney, String date, float xValue) {
+    public CandleEntry(float high, float low, float open, float close, float volume, float volumeMoney, String date, float xValue) {
         this.high = high;
         this.low = low;
         this.open = open;
@@ -58,7 +58,7 @@ public class CandleEntry {
         this.close = close;
     }
 
-    public int getVolume() {
+    public float getVolume() {
         return volume;
     }
 
@@ -66,7 +66,7 @@ public class CandleEntry {
         this.volume = volume;
     }
 
-    public int getVolumeMoney() {
+    public float getVolumeMoney() {
         return volumeMoney;
     }
